Chatbot/backend/requirements.txt
Flask==2.2.5
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.1
flask-cors==3.0.10



